+++ R/easystats.R
@@ -5,12 +5,28 @@ glance_easystats <- function(model, ...) {
   if (!check_dependency("performance")) {
     return(NULL)
   }
-  out <- performance::model_performance(model)
+  error_msg <- utils::capture.output(out <- performance::model_performance(model))
+
+  # lm model: include F-stat by default
+  if (isTRUE(class(model)[1] == "lm")) { # glm also inherits from lm
+    out$F <- attr(out, "r2")$F
+  }
+
   out <- insight::standardize_names(out, style="broom")
   mi <- insight::model_info(model)
+
+  # nobs
   if ("n_obs" %in% names(mi)) {
     out$nobs <- mi$n_obs
   }
+
+  # logLik
+  ll <- try(stats::logLik(model), silent=TRUE)
+  if (!inherits(ll, "try-error")) {
+    out$logLik <- as.numeric(ll[1])
+  }
+
+
   return(out)
 }